Oh joy: A grand compromise. Restore pynamesdef, w/o the import *
