+++ rope/refactor/importutils/module_imports.py
@@ -4,6 +4,7 @@ from rope.base import (
     ast,
     exceptions,
     pynames,
+    pynamesdef,
     utils,
 )
 from rope.refactor.importutils import actions, importinfo
@@ -36,7 +37,7 @@ class ModuleImports:
 
     def _get_all_star_list(self, pymodule):
         def _resolve_name(
-            name: Union[pynames.AssignedName, pynames.ImportedName]
+            name: Union[pynamesdef.AssignedName, pynames.ImportedName]
         ) -> List:
             while isinstance(name, pynames.ImportedName):
                 try:
@@ -45,7 +46,7 @@ class ModuleImports:
                     )
                 except exceptions.AttributeNotFoundError:
                     return []
-            assert isinstance(name, pynames.AssignedName)
+            assert isinstance(name, pynamesdef.AssignedName)
             return name.assignments
 
         result = set()
